DDF-1058 Simplify interface to eliminate builder pattern.  - Leave rest security where it was, secure the cookie, add method for insecure cookies.  - Instead of generating saml assertion, get x509 subject from the framework and use that.
