+++ rest/security-rest-cxfwrapper/src/main/java/org/codice/ddf/cxf/RestSecurity.java
@@ -20,13 +20,17 @@ import ddf.security.assertion.SecurityAssertion;
 import org.apache.cxf.common.util.Base64Utility;
 import org.apache.cxf.jaxrs.client.Client;
 import org.apache.cxf.rs.security.saml.DeflateEncoderDecoder;
+import org.apache.cxf.ws.security.tokenstore.SecurityToken;
 import org.apache.ws.security.WSSecurityException;
 import org.apache.ws.security.saml.ext.AssertionWrapper;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import javax.ws.rs.core.Cookie;
+import javax.ws.rs.core.NewCookie;
+import java.math.BigDecimal;
 import java.nio.charset.StandardCharsets;
+import java.util.Date;
 
 /**
  * Provides methods that help with securing RESTful (jaxrs) communications.
@@ -44,7 +48,25 @@ public final class RestSecurity {
      */
     public static void setSubjectOnClient(Subject subject, Client client) {
         if (subject != null) {
-            javax.ws.rs.core.Cookie cookie = createSamlCookie(subject);
+            javax.ws.rs.core.Cookie cookie = createSamlCookie(subject, true);
+            if (cookie == null) {
+                LOGGER.debug("SAML Cookie was null. Unable to set the cookie for the client.");
+                return;
+            }
+            client.cookie(cookie);
+        }
+    }
+
+    /**
+     * Sets a saml cookie without requiring ssl on the underlying client. This method
+     * only exists for compatibility with legacy or misconfigured systems, and is not
+     * recommended for use otherwise.
+     *
+     * @see #setSubjectOnClient(ddf.security.Subject, org.apache.cxf.jaxrs.client.Client)
+     */
+    public static void setUnsecuredSubjectOnClient(Subject subject, Client client) {
+        if (subject != null) {
+            javax.ws.rs.core.Cookie cookie = createSamlCookie(subject, false);
             if (cookie == null) {
                 LOGGER.debug("SAML Cookie was null. Unable to set the cookie for the client.");
                 return;
@@ -58,21 +80,29 @@ public final class RestSecurity {
      * a SAML assertion.
      *
      * @param subject - {@link ddf.security.Subject} to create the cookie from
+     * @param secure  - whether or not to require SSL for this cookie
      */
-    private static Cookie createSamlCookie(Subject subject) {
+    private static Cookie createSamlCookie(Subject subject, boolean secure) {
         Cookie cookie = null;
         org.w3c.dom.Element samlToken = null;
+        Date expires = null;
         try {
             for (Object principal : subject.getPrincipals().asList()) {
                 if (principal instanceof SecurityAssertion) {
-                    samlToken = ((SecurityAssertion) principal).getSecurityToken()
-                            .getToken();
+                    SecurityToken securityToken = ((SecurityAssertion) principal)
+                            .getSecurityToken();
+                    samlToken = securityToken.getToken();
+                    expires = securityToken.getExpires();
                 }
             }
             if (samlToken != null) {
-                cookie = new Cookie(SecurityConstants.SAML_COOKIE_NAME, encodeSaml(samlToken));
+                cookie = new NewCookie(
+                        new Cookie(SecurityConstants.SAML_COOKIE_NAME, encodeSaml(samlToken)), "",
+                        // gives us a checked exception for the cast
+                        new BigDecimal((expires.getTime() - new Date().getTime()) / 1000)
+                                .intValueExact(), secure).toCookie();
             }
-        } catch (WSSecurityException e) {
+        } catch (WSSecurityException | ArithmeticException e) {
             LOGGER.error("Unable to parse SAML assertion from subject.", e);
         }
         return cookie;
